Move use of funds components
Move components from main component directory to their own directory.
