+++ skins/laika/src/components/pages/use_of_funds/FundSection.vue
@@ -44,7 +44,7 @@ export default Vue.extend( {
 } );
 </script>
 <style lang="scss">
-@import "../scss/custom";
+@import "../../../scss/custom";
 
 .fund-text-inline {
 	display: flex;
